Dataset, to use PCA to improve upon the clustering and visualization done in assignment 2.
Problem Statement:
The marketing department for a credit card company wants to learn more about the different kinds of customers they have in order to develop a marketing strategy that targets each group separately. As a data scientist, your job is to find the best way to divide the customers into groups and you want to see if using PCA can help. In the end you will need to provide a description of each group to be used by the marketing strategists.